fix(server): Parse ASR_QPS_LIMIT to number
The `p-queue` library expects the `intervalCap` option to be a number, but environment variables are always read as strings. This caused a TypeError when the `ASR_QPS_LIMIT` environment variable was set.

This commit fixes the issue by parsing the environment variable to an integer using `parseInt()` before passing it to the `PQueue` constructor.